Utils_CLI_project

Function main

pub fn main()

Point d’entrée de l’application.

Cette fonction affiche un menu interactif permettant de choisir un utilitaire Linux parmi : ls, cat, cp, mv, rm, wc. L’utilisateur peut entrer le numéro correspondant ou quit pour quitter l’application.

L’application boucle tant que l’utilisateur ne choisit pas de quitter. Elle gère les entrées invalides en affichant un message d’erreur.

Algorithme

Boucle infinie jusqu’à ce que l’utilisateur tape quit.

Affiche le menu avec les différentes options disponibles.

Demande à l’utilisateur de saisir son choix.

Lit l’entrée utilisateur et supprime les espaces et retours à la ligne.

Vérifie si l’entrée est quit : si oui, quitte la boucle et termine le programme.

Sinon, effectue un match sur la saisie :

"1" : appelle la fonction ls::ls()

"2" : appelle la fonction cat::cat()

"3" : appelle la fonction cp::cp()

"4" : appelle la fonction mv::mv()

"5" : appelle la fonction rm::rm()

"6" : appelle la fonction wc::wc()

Autre : affiche un message d’erreur et redemande une saisie.

Affiche une ligne vide pour séparer les itérations.

Ce processus se répète jusqu’à ce que l’utilisateur décide de quitter.